requirements.txt
adal==1.2.1
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.4
asn1crypto==1.3.0
astroid==2.4.2
Babel==2.8.0
billiard==3.3.0.23
boto3==1.9.243
botocore==1.12.243
CairoSVG==2.4.2
celery==3.1.26.post2
certifi==2019.9.11
chardet==3.0.4
cryptography
cssselect2==0.3.0
defusedxml==0.6.0
distlib==0.3.1
dj-database-url
Django
django-allauth==0.40.0
django-anymail==7.0.0
django-background-tasks==1.2.5
django-celery==3.3.1
django-celery-beat==2.0.0
django-compat==1.0.15
django-cors-headers==3.1.1
django-extra-fields==2.0.5
django-filter==2.2.0
django-mail-templated==2.6.5
django-phonenumber-field==4.0.0
django-render-block==0.6
django-rest-auth==0.9.5
django-simple-history==2.8.0
django-storages==1.7.2
django-templated-email==2.3.0
django-timezone-field==4.0
django-weasyprint==1.0.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docutils==0.15.2
drf-extensions==0.5.0
filelock==3.0.12
gunicorn==19.9.0
html5lib==1.1
idna==2.8
importlib-metadata==1.6.0
isort==4.3.21
jmespath==0.9.4
kombu==3.0.37
lazy-object-proxy==1.4.3
Markdown==3.1
mccabe==0.6.1
oauthlib==3.1.0
onesignal-notifications==0.2.6
onesignal-sdk==1.1.0
pandas
pexpect==4.8.0
phonenumbers==8.12.6
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pylint-django==2.1.0
pylint-plugin-utils==0.6
Pyphen==0.9.5
python-crontab==2.4.1
python-dateutil==2.8.0
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.2
regex==2020.7.14
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.15.0
sqlparse==0.3.0
tinycss2==1.0.2
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.6
vine==1.3.0
virtualenv==20.0.27
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.0.1
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.3.6
zipp==3.1.0
